create folder for kits code, refactor imports to use @cmt and @test
