+++ src/cmakeProject.ts
@@ -23,34 +23,34 @@ import {
     ExecutableTarget,
     NoGeneratorError
 } from '@cmt/drivers/drivers';
-import { CTestDriver } from './ctest';
-import { CPackDriver } from './cpack';
-import { WorkflowDriver } from './workflow';
-import { CMakeBuildConsumer } from './diagnostics/build';
-import { CMakeOutputConsumer } from './diagnostics/cmake';
-import { FileDiagnostic, populateCollection } from './diagnostics/util';
-import { expandStrings, expandString, ExpansionOptions } from './expand';
-import { CMakeGenerator, Kit, SpecialKits } from './kit';
-import * as logging from './logging';
-import { fs } from './pr';
+import { CTestDriver } from '@cmt/ctest';
+import { CPackDriver } from '@cmt/cpack';
+import { WorkflowDriver } from '@cmt/workflow';
+import { CMakeBuildConsumer } from '@cmt/diagnostics/build';
+import { CMakeOutputConsumer } from '@cmt/diagnostics/cmake';
+import { FileDiagnostic, populateCollection } from '@cmt/diagnostics/util';
+import { expandStrings, expandString, ExpansionOptions } from '@cmt/expand';
+import { CMakeGenerator, Kit, SpecialKits } from '@cmt/kits/kit';
+import * as logging from '@cmt/logging';
+import { fs } from '@cmt/pr';
 import { buildCmdStr, DebuggerEnvironmentVariable, ExecutionResult, ExecutionOptions } from './proc';
-import { FireLate, Property } from './prop';
-import rollbar from './rollbar';
-import * as telemetry from './telemetry';
-import { VariantManager } from './variant';
+import { FireLate, Property } from '@cmt/prop';
+import rollbar from '@cmt/rollbar';
+import * as telemetry from '@cmt/telemetry';
+import { VariantManager } from '@cmt/variant';
 import * as nls from 'vscode-nls';
-import { ConfigurationWebview } from './cacheView';
-import { enableFullFeatureSet, extensionManager, updateFullFeatureSet, setContextAndStore } from './extension';
-import { CMakeCommunicationMode, ConfigurationReader, OptionConfig, UseCMakePresets, checkConfigureOverridesPresent } from './config';
-import * as preset from '@cmt/preset';
+import { ConfigurationWebview } from '@cmt/cacheView';
+import { enableFullFeatureSet, extensionManager, updateFullFeatureSet, setContextAndStore } from '@cmt/extension';
+import { CMakeCommunicationMode, ConfigurationReader, OptionConfig, UseCMakePresets, checkConfigureOverridesPresent } from '@cmt/config';
+import * as preset from '@cmt/presets/preset';
 import * as util from '@cmt/util';
-import { Environment, EnvironmentUtils } from './environmentVariables';
-import { KitsController } from './kitsController';
-import { PresetsController } from './presetsController';
-import paths from './paths';
-import { ProjectController } from './projectController';
+import { Environment, EnvironmentUtils } from '@cmt/environmentVariables';
+import { KitsController } from '@cmt/kits/kitsController';
+import { PresetsController } from '@cmt/presets/presetsController';
+import paths from '@cmt/paths';
+import { ProjectController } from '@cmt/projectController';
 import { MessageItem } from 'vscode';
-import { DebugTrackerFactory, DebuggerInformation, getDebuggerPipeName } from './debug/debuggerConfigureDriver';
+import { DebugTrackerFactory, DebuggerInformation, getDebuggerPipeName } from '@cmt/debug/debuggerConfigureDriver';
 import { ConfigurationType } from 'vscode-cmake-tools';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
