+++ src/diagnostics/diab.ts
@@ -4,7 +4,7 @@
 
 import * as vscode from 'vscode';
 
-import { oneLess, RawDiagnosticParser, FeedLineResult } from './util';
+import { oneLess, RawDiagnosticParser, FeedLineResult } from '@cmt/diagnostics/util';
 
 export const REGEX = /^\"(.*)\",\s+(?:line\s+(\d+):\s+)?(info|warning|(?:|fatal |catastrophic )error)\s+\((.*)\):\s+(.*)$/;
