+++ src/diagnostics/iar.ts
@@ -4,7 +4,7 @@
 
 import * as vscode from 'vscode';
 
-import { oneLess, RawDiagnosticParser, FeedLineResult, RawDiagnostic } from './util';
+import { oneLess, RawDiagnosticParser, FeedLineResult, RawDiagnostic } from '@cmt/diagnostics/util';
 
 const CODE_REGEX = /^\"(?<file>.*)\",(?<line>\d+)\s+(?<severity>[A-Za-z ]+)\[(?<code>[A-Za-z]+[0-9]+)\]:(?<message_start>.*)$/;
