+++ src/diagnostics/msvc.ts
@@ -4,7 +4,7 @@
 
 import * as vscode from 'vscode';
 
-import { oneLess, RawDiagnosticParser, FeedLineResult } from './util';
+import { oneLess, RawDiagnosticParser, FeedLineResult } from '@cmt/diagnostics/util';
 
 export const REGEX = /^\s*(\d+>)?\s*([^\s>].*)\((\d+|\d+,\d+|\d+,\d+,\d+,\d+)\)\s*:\s+((?:fatal )?error|warning|info)\s*(\w{1,2}\d+)?\s*:\s*(.*)$/;
