+++ src/drivers/drivers.ts
@@ -1,6 +1,6 @@
-export * from './cmakeDriver';
-export * from './cmakeFileApi';
-export * from './cmakeFileApiDriver';
-export * from './cmakeLegacyDriver';
-export * from './cmakeServerClient';
-export * from './cmakeServerDriver';
+export * from '@cmt/drivers/cmakeDriver';
+export * from '@cmt/drivers/cmakeFileApi';
+export * from '@cmt/drivers/cmakeFileApiDriver';
+export * from '@cmt/drivers/cmakeLegacyDriver';
+export * from '@cmt/drivers/cmakeServerClient';
+export * from '@cmt/drivers/cmakeServerDriver';
