+++ src/kits/kit.ts
@@ -8,21 +8,21 @@ import * as util from '@cmt/util';
 import * as json5 from 'json5';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import * as kitsController from '@cmt/kitsController';
+import * as kitsController from '@cmt/kits/kitsController';
 
-import CMakeProject from './cmakeProject';
-import * as expand from './expand';
-import { VSInstallation, vsInstallations, getHostTargetArchString, varsForVSInstallation, generatorPlatformFromVSArch } from './installs/visualStudio';
-import * as logging from './logging';
-import paths, { PathWithTrust } from './paths';
-import { fs } from './pr';
-import * as proc from './proc';
-import { loadSchema } from './schema';
-import { TargetTriple, findTargetTriple, parseTargetTriple, computeTargetTriple } from './triple';
-import { compare, dropNulls, Ordering, versionLess } from './util';
+import CMakeProject from '@cmt/cmakeProject';
+import * as expand from '@cmt/expand';
+import { VSInstallation, vsInstallations, getHostTargetArchString, varsForVSInstallation, generatorPlatformFromVSArch } from '@cmt/installs/visualStudio';
+import * as logging from '@cmt/logging';
+import paths, { PathWithTrust } from '@cmt/paths';
+import { fs } from '@cmt/pr';
+import * as proc from '@cmt/proc';
+import { loadSchema } from "@cmt/schema";
+import { TargetTriple, findTargetTriple, parseTargetTriple, computeTargetTriple } from '@cmt/triple';
+import { compare, dropNulls, Ordering, versionLess } from "@cmt/util";
 import * as nls from 'vscode-nls';
-import { Environment, EnvironmentUtils } from './environmentVariables';
-import { getCMakeExecutableInformation } from './cmake/cmakeExecutable';
+import { Environment, EnvironmentUtils } from "@cmt/environmentVariables";
+import { getCMakeExecutableInformation } from "@cmt/cmake/cmakeExecutable";
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
