+++ src/presets/preset.ts
@@ -10,8 +10,8 @@ import { execute } from '@cmt/proc';
 import { errorHandlerHelper, expandString, ExpansionErrorHandler, ExpansionOptions } from '@cmt/expand';
 import paths from '@cmt/paths';
 import { compareVersions, VSInstallation, vsInstallations, enumerateMsvcToolsets, varsForVSInstallation, getVcVarsBatScript } from '@cmt/installs/visualStudio';
-import { EnvironmentUtils, EnvironmentWithNull } from './environmentVariables';
-import { defaultNumJobs, UseVsDeveloperEnvironment } from './config';
+import { EnvironmentUtils, EnvironmentWithNull } from '@cmt/environmentVariables';
+import { defaultNumJobs, UseVsDeveloperEnvironment } from '@cmt/config';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
