+++ src/presets/presetsController.ts
@@ -7,13 +7,13 @@ import * as lodash from "lodash";
 import { CMakeProject, ConfigureTrigger, ConfigureType } from '@cmt/cmakeProject';
 import * as logging from '@cmt/logging';
 import { fs } from '@cmt/pr';
-import * as preset from '@cmt/preset';
+import * as preset from '@cmt/presets/preset';
 import * as util from '@cmt/util';
 import rollbar from '@cmt/rollbar';
 import { expandString, ExpansionErrorHandler, ExpansionOptions, MinimalPresetContextVars } from '@cmt/expand';
 import paths from '@cmt/paths';
-import { KitsController } from '@cmt/kitsController';
-import { descriptionForKit, Kit, SpecialKits } from '@cmt/kit';
+import { KitsController } from '@cmt/kits/kitsController';
+import { descriptionForKit, Kit, SpecialKits } from '@cmt/kits/kit';
 import { getHostTargetArchString } from '@cmt/installs/visualStudio';
 import { loadSchema } from '@cmt/schema';
 import json5 = require('json5');
