+++ src/projectController.ts
@@ -9,17 +9,17 @@ import * as logging from '@cmt/logging';
 import CMakeProject from '@cmt/cmakeProject';
 import rollbar from '@cmt/rollbar';
 import { disposeAll, DummyDisposable } from '@cmt/util';
-import { ConfigurationReader, OptionConfig } from './config';
-import { CMakeDriver } from './drivers/drivers';
-import { DirectoryContext } from './workspace';
-import { StateManager } from './state';
-import { getStatusBar } from './extension';
-import * as telemetry from './telemetry';
-import { StatusBar } from './status';
-import { FireNow } from './prop';
-import { setContextAndStore } from './extension';
-import * as ext from './extension';
-import { ProjectStatus } from './projectStatus';
+import { ConfigurationReader, OptionConfig } from '@cmt/config';
+import { CMakeDriver } from '@cmt/drivers/drivers';
+import { DirectoryContext } from '@cmt/workspace';
+import { StateManager } from '@cmt/state';
+import { getStatusBar } from '@cmt/extension';
+import * as telemetry from '@cmt/telemetry';
+import { StatusBar } from '@cmt/status';
+import { FireNow } from '@cmt/prop';
+import { setContextAndStore } from '@cmt/extension';
+import * as ext from '@cmt/extension';
+import { ProjectStatus } from '@cmt/projectStatus';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
