+++ test/end-to-end-tests/successful-build/test/scan_kits.test.ts
@@ -1,4 +1,4 @@
-import { scanForKits } from '@cmt/kit';
+import { scanForKits } from '@cmt/kits/kit';
 import { fs } from '@cmt/pr';
 import { expect } from '@test/util';
