+++ test/helpers/cmake/project-root-helper.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { BuildDirectoryHelper } from './build-directory-helper';
+import { BuildDirectoryHelper } from '@test/helpers/cmake/build-directory-helper';
 
 export class ProjectRootHelper {
     private readonly _locationOfThisClassFile: string = __dirname;
