+++ test/unit-tests/driver/cmakeServerDriver.test.ts
@@ -1,9 +1,9 @@
 import { CMakeExecutable } from '@cmt/cmake/cmakeExecutable';
 import { ConfigurationReader } from '@cmt/config';
 import { CMakeServerDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/drivers';
-import { CMakeGenerator, Kit } from '@cmt/kit';
-import { makeCodeModelDriverTestsuite } from './driver-codemodel-tests';
-import { makeDriverTestsuite } from './driver-test';
+import { CMakeGenerator, Kit } from '@cmt/kits/kit';
+import { makeCodeModelDriverTestsuite } from '@test/unit-tests/driver/driver-codemodel-tests';
+import { makeDriverTestsuite } from '@test/unit-tests/driver/driver-test';
 
 async function cmakeServerDriverFactory(cmake: CMakeExecutable, config: ConfigurationReader, kit: Kit | null, workspaceFolder: string, preconditionHandler: CMakePreconditionProblemSolver, preferredGenerators: CMakeGenerator[]) {
     const d: CMakeServerDriver = await CMakeServerDriver.create(cmake, config, workspaceFolder || "", false, false, kit, null, null, null, null, null, workspaceFolder, preconditionHandler, preferredGenerators);
